drop Beacon-Age header
Current implementations do not defer or coalesce beacon requests, and
the definition of Beacon-Age allows it to be omitted when age is 0. As
such, this is backwards compatible change.

The right place to enable this is probably in Fetch API:
https://github.com/whatwg/fetch/issues/184

Assuming above is resolved, Beacon API would simply set a flag to
indicate to fetch that the request is delay tolerant. Fetch would handle
the coalescing logic and append the ~Request-Age header. This would
allow different request types to be batched together (e.g. background
sync, beacons, and so on).