Set the keepalive flag when extracting the body
Setting the keepalive flag is necessary so that the fetch algorithm will
throw a TypeError if a ReadableStream is passed in. See
https://crbug.com/973659.

The requirement to throw is already tested by the web platform tests:
https://wpt.fyi/results/beacon/beacon-readablestream.window.html.

Also rephrase the extract step to match the way it is used in the Fetch
Standard.